+++ lib/cpp/src/thrift/transport/TTransportUtils.cpp
@@ -111,7 +111,7 @@ TPipedFileReaderTransport::TPipedFileReaderTransport(
 TPipedFileReaderTransport::~TPipedFileReaderTransport() {
 }
 
-bool TPipedFileReaderTransport::isOpen() {
+bool TPipedFileReaderTransport::isOpen() const {
   return TPipedTransport::isOpen();
 }